requirements.txt
fastapi
uvicorn
asyncpg
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pyjwt